pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taf-code"
version = "1.0.0"
description = "AI-powered coding assistant for your terminal, powered by Gemini"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["gemini", "ai", "coding", "assistant", "cli", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Terminals",
]
dependencies = [
    "google-genai>=1.0.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "pygments>=2.0.0",
    "pathspec>=1.0.0",
]

[project.scripts]
taf-code = "taf_code.main:main"
taf = "taf_code.main:main"

[project.urls]
Homepage = "https://github.com/gemini-code/gemini-code"
Issues = "https://github.com/gemini-code/gemini-code/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["taf_code*"]
